+++ server/controllers/cohortsController.js
@@ -1,12 +1,24 @@
 const express = require("express");
+const session = require("express-session");
 const router = express.Router();
 const Cohort = require("../models/cohort");
 const seed = require("../seed/seedCohort");
 
-router.get("/seed", seed);
+// session
+
+router.use(
+  session({
+    secret: process.env.SECRET,
+    resave: false,
+    saveUninitialized: true,
+    // cookie: { secure: true },
+  })
+);
+
+router.get("/seed", seed); // DELETE!
 
 router.get("/", async (req, res) => {
-  //? return [ list of cohortss]
+  //? return [ list of cohorts]
   try {
     const cohorts = await Cohort.find().exec();
     res.json(cohorts);
@@ -16,6 +28,11 @@ router.get("/", async (req, res) => {
 });
 
 router.post("/", async (req, res) => {
+  // Check for the presence of session data
+  if (!req.session.username) {
+    res.status(401).send("Unauthorized");
+    return;
+  }
   try {
     const cohort = await Cohort.create(req.body);
     res.status(201).json(cohort);
@@ -25,8 +42,12 @@ router.post("/", async (req, res) => {
 });
 
 router.delete("/:id", async (req, res) => {
+  // Check for the presence of session data
+  if (!req.session.username) {
+    res.status(401).send("Unauthorized");
+    return;
+  }
   const { id } = req.params;
-
   try {
     const deletedCohort = await Cohort.findByIdAndRemove(id);
     res.status(200).send(deletedCohort);
@@ -46,6 +67,11 @@ router.get("/:id", async (req, res) => {
 });
 
 router.put("/:id", async (req, res) => {
+  // Check for the presence of session data
+  if (!req.session.username) {
+    res.status(401).send("Unauthorized");
+    return;
+  }
   const { id } = req.params;
   try {
     const updatedCohort = await Cohort.findByIdAndUpdate(id, req.body, {
